Move STACObjectType from pystac.serialization to pystac.stac_object
Also import it into the top level package, so users can use as `pystac.STACObjectType`.
