+++ pystac/serialization/__init__.py
@@ -3,8 +3,10 @@ from pystac import (Catalog, Collection, Item, Extensions)
 
 from pystac.serialization.identify import (STACObjectType, STACJSONDescription, STACVersionRange,
                                            STACVersionID, identify_stac_object,
-                                           identify_stac_object_type)
+from pystac import (Catalog, Collection, Item, Extensions, STACObjectType)
 
+from pystac.serialization.identify import (STACJSONDescription, STACVersionRange, STACVersionID,
+                                           identify_stac_object, identify_stac_object_type)
 from pystac.serialization.common_properties import merge_common_properties
 from pystac.serialization.migrate import migrate_to_latest
